# docs-site/docs/api/openapi/core-full.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AstroEngine API",
    "version": "1.0"
  },
  "paths": {
...
    "/v1/natals": {
      "get": {
        "tags": [
          "natals"
        ],
        "summary": "List stored natal charts.",
        "description": "Return a paginated view of available natal records.",
        "operationId": "listNatals",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "1-indexed page number.",
              "default": 1,
              "title": "Page"
            },
            "description": "1-indexed page number."
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Number of items to include in the response page (max 100).",
              "default": 25,
              "title": "Page Size"
            },
            "description": "Number of items to include in the response page (max 100)."
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of natals.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NatalCollection"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/natals/{natal_id}": {
      "get": {
        "tags": [
          "natals"
        ],
        "summary": "Fetch a single natal chart.",
        "operationId": "getNatal",
        "parameters": [
          {
            "name": "natal_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Identifier returned from the list endpoint.",
              "title": "Natal Id"
            },
            "description": "Identifier returned from the list endpoint."
          }
        ],
        "responses": {
          "200": {
            "description": "The requested natal chart.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NatalRecord"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "natals"
        ],
        "summary": "Create or replace a natal chart.",
        "operationId": "upsertNatal",
        "parameters": [
          {
            "name": "natal_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Identifier to assign to the natal chart.",
              "title": "Natal Id"
            },
            "description": "Identifier to assign to the natal chart."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/astroengine__api__routers__natals__NatalPayload"
              },
              "examples": {
                "apollo11": {
                  "summary": "Apollo 11 launch data",
                  "value": {
                    "name": "Apollo 11 Launch",
                    "utc": "1969-07-16T13:32:00Z",
                    "lat": 28.60839,
                    "lon": -80.60433,
                    "tz": "UTC",
                    "place": "Launch Complex 39A, Kennedy Space Center"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The persisted natal chart.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NatalRecord"
                }
              }
            }
          },
          "201": {
            "description": "The persisted natal chart."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorEnvelope": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "description": "Machine readable error code."
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Human friendly summary of the error."
          },
          "details": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Details",
            "description": "Optional structured details that expand on the error."
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorEnvelope",
        "description": "Standardized error payload returned by the public API."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "NatalCollection": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/NatalRecord"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "page_size"
        ],
        "title": "NatalCollection",
        "description": "Paginated collection of natals returned by list endpoints.",
        "example": {
          "items": [
            {
              "lat": 28.60839,
              "lon": -80.60433,
              "name": "Apollo 11 Launch",
              "natal_id": "apollo-11-launch",
              "place": "Launch Complex 39A, Kennedy Space Center",
              "tz": "UTC",
              "utc": "1969-07-16T13:32:00Z"
            }
          ],
          "page": 1,
          "page_size": 25,
          "total": 1
        }
      },
      "NatalRecord": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name",
            "description": "Display name for the chart owner."
          },
          "utc": {
            "type": "string",
            "title": "Utc",
            "description": "Moment of birth in ISO-8601 UTC."
          },
          "lat": {
            "type": "number",
            "title": "Lat",
            "description": "Latitude in decimal degrees."
          },
          "lon": {
            "type": "number",
            "title": "Lon",
            "description": "Longitude in decimal degrees."
          },
          "tz": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tz",
            "description": "Original timezone identifier, if known."
          },
          "place": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Place",
            "description": "Birth location description."
          },
          "natal_id": {
            "type": "string",
            "title": "Natal Id",
            "description": "Stable identifier for the natal record."
          }
        },
        "type": "object",
        "required": [
          "utc",
          "lat",
          "lon",
          "natal_id"
        ],
        "title": "NatalRecord",
        "description": "Representation of a stored natal chart.",
        "examples": [
          {
            "lat": 28.60839,
            "lon": -80.60433,
            "name": "Apollo 11 Launch",
            "place": "Launch Complex 39A, Kennedy Space Center",
            "tz": "UTC",
            "utc": "1969-07-16T13:32:00Z"
          }
        ]
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "astroengine__api__routers__natals__NatalPayload": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name",
            "description": "Display name for the chart owner."
          },
          "utc": {
            "type": "string",
            "format": "date-time",
            "title": "Utc",
            "description": "Moment of birth in ISO-8601 UTC."
          },
          "lat": {
            "type": "number",
            "title": "Lat",
            "description": "Latitude in decimal degrees."
          },
          "lon": {
            "type": "number",
            "title": "Lon",
            "description": "Longitude in decimal degrees."
          },
          "tz": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tz",
            "description": "Original timezone identifier, if known."
          },
          "place": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Place",
            "description": "Birth location description."
          }
        },
        "type": "object",
        "required": [
          "utc",
          "lat",
          "lon"
        ],
        "title": "NatalPayload",
        "description": "Incoming payload used to create or update a natal record.",
        "examples": [
          {
            "lat": 28.60839,
            "lon": -80.60433,
            "name": "Apollo 11 Launch",
            "place": "Launch Complex 39A, Kennedy Space Center",
            "tz": "UTC",
            "utc": "1969-07-16T13:32:00Z"
          }
        ]
      },
...
    }
  },
...
}